requirements.txt
asgiref==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.3
django-import-export==2.4.0
django-multiselectfield==0.1.12
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.0.1
psycopg2==2.8.6
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.2.0
xlwt==1.3.0
